Integrate python frontend testing in CMake
